Make add goal and add experience into separate servlet
