+++ src/main/java/com/google/sps/servlets/GoalServlet.java
@@ -58,28 +58,4 @@ public class GoalServlet extends HttpServlet {
         response.getWriter().println(json);
         return;
     }
-
-    @Override
-    public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        //Set default value to -1 
-        String studentID = Optional.ofNullable((String)request.getSession(false).getAttribute("userId")).orElse("-1");
-        String goal = request.getParameter("goal");
-
-        if(studentID.equals("-1")) {
-            response.setContentType("application/json");
-            response.getWriter().println("{\"error\": \"There was an error adding goal.\"}");
-            return;
-        }
-
-        Goal newGoal = new Goal(studentID, goal);
-
-        // Add goal
-        datastore.addGoal(newGoal);
-
-        String json = new Gson().toJson(datastore.getGoalsByStudent(studentID));
-        response.setContentType("application/json;");
-        response.getWriter().println(json);
-        response.sendRedirect("/progress.html?studentID=" + studentID);
-        return;
-    }
 }
