+++ src/main/webapp/WEB-INF/web.xml
@@ -9,19 +9,21 @@
     </filter>
     <filter-mapping>
         <filter-name>SessionFilter</filter-name>
-        <url-pattern>add-availability</url-pattern>
+        <url-pattern>add-goal</url-pattern>
+        <url-pattern>add-experience</url-pattern>
         <url-pattern>confirmation</url-pattern>
         <url-pattern>delete-availability</url-pattern>
         <url-pattern>delete-experience</url-pattern>
         <url-pattern>delete-goal</url-pattern>
         <url-pattern>delete-tutor-session</url-pattern>
         <url-pattern>history</url-pattern>
+        <url-pattern>manage-availability</url-pattern>
         <url-pattern>my-students</url-pattern>
         <url-pattern>rating</url-pattern>
         <url-pattern>registration</url-pattern>
         <url-pattern>scheduling</url-pattern>
     </filter-mapping>
-    <security-constraint>
+    <!-- <security-constraint>
         <web-resource-collection>
             <url-pattern>/history.html</url-pattern>
         </web-resource-collection>
@@ -68,5 +70,5 @@
         <auth-constraint>
             <role-name>*</role-name>
         </auth-constraint>
-    </security-constraint>
+    </security-constraint> -->
 </web-app>
