+++ src/test/java/com/google/sps/AddExperienceTest.java
@@ -32,7 +32,7 @@ import org.junit.After;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 import static org.mockito.Mockito.*;
-import com.google.sps.servlets.ExperienceServlet;
+import com.google.sps.servlets.AddExperienceServlet;
 import com.google.sps.data.SampleData;
 import com.google.sps.data.TimeRange;
 import com.google.sps.data.Tutor;
@@ -45,16 +45,16 @@ import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
 
 @RunWith(JUnit4.class)
-public final class ExperienceTest {
+public final class AddExperienceTest {
     private final LocalServiceTestHelper helper =  new LocalServiceTestHelper(new LocalDatastoreServiceTestConfig()); 
 
-    private ExperienceServlet servlet;
+    private AddExperienceServlet servlet;
 
     @Before
     public void setUp() {
         helper.setUp();
 
-        servlet = new ExperienceServlet();
+        servlet = new AddExperienceServlet();
         servlet.init();
 
         SampleData sample = new SampleData();
@@ -66,82 +66,54 @@ public final class ExperienceTest {
     }
 
     @Test
-    public void testDoGetWithNoExperiences() throws Exception {
+    public void testDoPost() throws Exception {
         HttpServletRequest request = mock(HttpServletRequest.class);       
         HttpServletResponse response = mock(HttpServletResponse.class);
 
-        when(request.getParameter("studentID")).thenReturn("123");
+        TestUtilities.setSessionId(request, "123");   
+        when(request.getParameter("experience")).thenReturn("testing");
 
         StringWriter stringWriter = new StringWriter();
         PrintWriter writer = new PrintWriter(stringWriter);
         when(response.getWriter()).thenReturn(writer);
         when(request.getContentType()).thenReturn("application/json");
 
-        servlet.doGet(request, response);
+        servlet.doPost(request, response);
 
-        verify(request, times(1)).getParameter("studentID");
-
-        String expected = "[]";
-      
-        writer.flush();
-
-        // No experiences should have been returned
-        Assert.assertTrue(stringWriter.toString().contains(expected));
-    }
-
-    @Test
-    public void testDoGetWithExperiences() throws Exception {
-        HttpServletRequest request = mock(HttpServletRequest.class);       
-        HttpServletResponse response = mock(HttpServletResponse.class);
-
-        // Adds experience entity to the local datastore
-        DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-        Entity experienceEntity = new Entity("Experience");
-        experienceEntity.setProperty("studentID", "123");
-        experienceEntity.setProperty("experience", "testing");
-        datastore.put(experienceEntity);
-
-        when(request.getParameter("studentID")).thenReturn("123");
-
-        StringWriter stringWriter = new StringWriter();
-        PrintWriter writer = new PrintWriter(stringWriter);
-        when(response.getWriter()).thenReturn(writer);
-        when(request.getContentType()).thenReturn("application/json");
-
-        servlet.doGet(request, response);
-
-        verify(request, times(1)).getParameter("studentID");
+        verify(request, times(1)).getParameter("experience");
 
         String expected = new Gson()
                             .toJson(new ArrayList<Experience> (Arrays.asList(new Experience("123", "testing", 1))));
       
         writer.flush();
 
-        // The correct list of experiences should have been returned
+        // New experience should have been added
         Assert.assertTrue(stringWriter.toString().contains(expected));
     }
 
     @Test
-    public void testDoGetWithError() throws Exception {
+    public void testDoPostWithError() throws Exception {
         HttpServletRequest request = mock(HttpServletRequest.class);       
         HttpServletResponse response = mock(HttpServletResponse.class);
 
-        when(request.getParameter("studentID")).thenReturn(null);
+        TestUtilities.setSessionId(request, null);           
+        when(request.getParameter("experience")).thenReturn("testing");
 
         StringWriter stringWriter = new StringWriter();
         PrintWriter writer = new PrintWriter(stringWriter);
         when(response.getWriter()).thenReturn(writer);
         when(request.getContentType()).thenReturn("application/json");
 
-        servlet.doGet(request, response);
+        servlet.doPost(request, response);
 
-        verify(request, times(1)).getParameter("studentID");
+        verify(request, times(1)).getParameter("experience");
 
-        String expected = "{\"error\": \"There was an error getting experiences.\"}";
+        String expected = "{\"error\": \"There was an error adding experience.\"}";
       
         writer.flush();
 
         // Error should have been returned
         Assert.assertTrue(stringWriter.toString().contains(expected));
     }
+
 }
