+++ src/test/java/com/google/sps/AddGoalTest.java
@@ -32,11 +32,11 @@ import org.junit.After;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 import static org.mockito.Mockito.*;
-import com.google.sps.servlets.ExperienceServlet;
+import com.google.sps.servlets.AddGoalServlet;
 import com.google.sps.data.SampleData;
 import com.google.sps.data.TimeRange;
 import com.google.sps.data.Tutor;
-import com.google.sps.data.Experience;
+import com.google.sps.data.Goal;
 import com.google.gson.Gson;
 import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
 import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
@@ -45,16 +45,16 @@ import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
 
 @RunWith(JUnit4.class)
-public final class ExperienceTest {
+public final class AddGoalTest {
     private final LocalServiceTestHelper helper =  new LocalServiceTestHelper(new LocalDatastoreServiceTestConfig()); 
 
-    private ExperienceServlet servlet;
+    private AddGoalServlet servlet;
 
     @Before
     public void setUp() {
         helper.setUp();
 
-        servlet = new ExperienceServlet();
+        servlet = new AddGoalServlet();
         servlet.init();
 
         SampleData sample = new SampleData();
@@ -66,82 +66,54 @@ public final class ExperienceTest {
     }
 
     @Test
-    public void testDoGetWithNoExperiences() throws Exception {
+    public void testDoPost() throws Exception {
         HttpServletRequest request = mock(HttpServletRequest.class);       
         HttpServletResponse response = mock(HttpServletResponse.class);
 
-        when(request.getParameter("studentID")).thenReturn("123");
+        TestUtilities.setSessionId(request, "123");   
+        when(request.getParameter("goal")).thenReturn("testing");
 
         StringWriter stringWriter = new StringWriter();
         PrintWriter writer = new PrintWriter(stringWriter);
         when(response.getWriter()).thenReturn(writer);
         when(request.getContentType()).thenReturn("application/json");
 
-        servlet.doGet(request, response);
+        servlet.doPost(request, response);
 
-        verify(request, times(1)).getParameter("studentID");
-
-        String expected = "[]";
-      
-        writer.flush();
-
-        // No experiences should have been returned
-        Assert.assertTrue(stringWriter.toString().contains(expected));
-    }
-
-    @Test
-    public void testDoGetWithExperiences() throws Exception {
-        HttpServletRequest request = mock(HttpServletRequest.class);       
-        HttpServletResponse response = mock(HttpServletResponse.class);
-
-        // Adds experience entity to the local datastore
-        DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-        Entity experienceEntity = new Entity("Experience");
-        experienceEntity.setProperty("studentID", "123");
-        experienceEntity.setProperty("experience", "testing");
-        datastore.put(experienceEntity);
-
-        when(request.getParameter("studentID")).thenReturn("123");
-
-        StringWriter stringWriter = new StringWriter();
-        PrintWriter writer = new PrintWriter(stringWriter);
-        when(response.getWriter()).thenReturn(writer);
-        when(request.getContentType()).thenReturn("application/json");
-
-        servlet.doGet(request, response);
-
-        verify(request, times(1)).getParameter("studentID");
+        verify(request, times(1)).getParameter("goal");
 
         String expected = new Gson()
-                            .toJson(new ArrayList<Experience> (Arrays.asList(new Experience("123", "testing", 1))));
+                            .toJson(new ArrayList<Goal> (Arrays.asList(new Goal("123", "testing", 1))));
       
         writer.flush();
 
-        // The correct list of experiences should have been returned
+        // New goal should have been added
         Assert.assertTrue(stringWriter.toString().contains(expected));
     }
 
     @Test
-    public void testDoGetWithError() throws Exception {
+    public void testDoPostWithError() throws Exception {
         HttpServletRequest request = mock(HttpServletRequest.class);       
         HttpServletResponse response = mock(HttpServletResponse.class);
 
-        when(request.getParameter("studentID")).thenReturn(null);
+        TestUtilities.setSessionId(request, null);           
+        when(request.getParameter("goal")).thenReturn("goal");
 
         StringWriter stringWriter = new StringWriter();
         PrintWriter writer = new PrintWriter(stringWriter);
         when(response.getWriter()).thenReturn(writer);
         when(request.getContentType()).thenReturn("application/json");
 
-        servlet.doGet(request, response);
+        servlet.doPost(request, response);
 
-        verify(request, times(1)).getParameter("studentID");
+        verify(request, times(1)).getParameter("goal");
 
-        String expected = "{\"error\": \"There was an error getting experiences.\"}";
+        String expected = "{\"error\": \"There was an error adding goal.\"}";
       
         writer.flush();
 
         // Error should have been returned
         Assert.assertTrue(stringWriter.toString().contains(expected));
     }
+
 }
