+++ src/test/java/com/google/sps/GoalTest.java
@@ -144,56 +144,4 @@ public final class GoalTest {
         // Error should have been returned
         Assert.assertTrue(stringWriter.toString().contains(expected));
     }
-
-    @Test
-    public void testDoPost() throws Exception {
-        HttpServletRequest request = mock(HttpServletRequest.class);       
-        HttpServletResponse response = mock(HttpServletResponse.class);
-
-        TestUtilities.setSessionId(request, "123");   
-        when(request.getParameter("goal")).thenReturn("testing");
-
-        StringWriter stringWriter = new StringWriter();
-        PrintWriter writer = new PrintWriter(stringWriter);
-        when(response.getWriter()).thenReturn(writer);
-        when(request.getContentType()).thenReturn("application/json");
-
-        servlet.doPost(request, response);
-
-        verify(request, times(1)).getParameter("goal");
-
-        String expected = new Gson()
-                            .toJson(new ArrayList<Goal> (Arrays.asList(new Goal("123", "testing", 1))));
-      
-        writer.flush();
-
-        // New goal should have been added
-        Assert.assertTrue(stringWriter.toString().contains(expected));
-    }
-
-    @Test
-    public void testDoPostWithError() throws Exception {
-        HttpServletRequest request = mock(HttpServletRequest.class);       
-        HttpServletResponse response = mock(HttpServletResponse.class);
-
-        TestUtilities.setSessionId(request, null);           
-        when(request.getParameter("goal")).thenReturn("goal");
-
-        StringWriter stringWriter = new StringWriter();
-        PrintWriter writer = new PrintWriter(stringWriter);
-        when(response.getWriter()).thenReturn(writer);
-        when(request.getContentType()).thenReturn("application/json");
-
-        servlet.doPost(request, response);
-
-        verify(request, times(1)).getParameter("goal");
-
-        String expected = "{\"error\": \"There was an error adding goal.\"}";
-      
-        writer.flush();
-
-        // Error should have been returned
-        Assert.assertTrue(stringWriter.toString().contains(expected));
-    }
-
 }
